Split GlobalTab into focused sub-components (closes #127)
Extracts two self-contained sections from the 1261-line GlobalTab:

- GlobalRemindersSection.tsx (189 lines) — moved from inline definition
  in GlobalTab into its own file; no logic changes
- GlobalLogSection.tsx (151 lines) — extracted log state, effects,
  and handlers (handleLogDelete, cancelLog, handleLogSubmit) from
  GlobalTab; props: { contact }

GlobalTab.tsx drops from 1261 → 944 lines. Each extracted component
owns its own IPC calls and state; the parent composes them with a
single contact prop.
